requirements.txt
torch
torchaudio
transformers
pinecone-client
python-dotenv
